server/requirements.txt
absl-py==1.4.0
accelerate==0.18.0
appdirs==1.4.4
astunparse==1.6.3
audioread==3.0.0
boto3==1.26.106
botocore==1.29.106
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
diffusers==0.14.0
filelock==3.10.7
Flask==2.0.1
Flask-Cors==3.0.10
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
huggingface==0.0.1
huggingface-hub==0.13.3
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
keras==2.12.0
lazy_loader==0.2
libclang==16.0.0
librosa==0.10.0.post2
llvmlite==0.39.1
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.0.4
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
pooch==1.6.0
protobuf==4.22.1
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.4
sympy==1.11.1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
tensorflow-macos==2.12.0
tensorflow-metal==0.8.0
termcolor==2.2.0
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.27.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uuid==1.30
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.15.0
